+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.1641 (2026-04-30)
+
+### Features
+- harden PR attachment contract (#1908)
+
+## 0.1.1640 (2026-04-30)
+
+### Fixes
+- surface unresolved pr branches
+
 ## 0.1.1638 (2026-04-30)
 
 ### Fixes